+++ backend/tests/test_library.c
@@ -84,7 +84,7 @@ static void test_checks(void)
         if (data[i].dot_size != -1) {
             symbol->dot_size = data[i].dot_size;
         }
-        int length = data[i].length == -1 ? strlen(data[i].data) : data[i].length;
+        int length = data[i].length == -1 ? (int) strlen(data[i].data) : data[i].length;
 
         ret = ZBarcode_Encode(symbol, data[i].data, length);
         assert_equal(ret, data[i].ret, "i:%d ZBarcode_Encode(%d) ret %d != %d (%s)\n", i, data[i].symbology, ret, data[i].ret, symbol->errtxt);
